requirements.txt
fastmcp>=2.0
mysql-connector-python>=8.0.0
python-levenshtein>=0.25.0
uvicorn>=0.24.0
starlette>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0